Move workspace into a separate file
Also move impl From<ProjectInfo> for Project into project.rs in order to not export workspace outside the project.
