+++ cli/src/project.rs
@@ -5,103 +5,25 @@ mod parse;
 
 /// Runtime templates for different workloads
 mod templates;
+mod workspace;
 
 use crate::api::client::Client;
-use crate::api::projects::Kvdb;
+use crate::api::projects::{Kvdb, ProjectInfo};
 use crate::api::request::Validate;
 use crate::api::stack;
 use crate::config::deploy::DeployConfig;
 use crate::envs::Envs;
 use crate::error::Error;
 use crate::function::Function;
+use crate::project::workspace::Workspace;
 use crate::secrets::Secrets;
 use cache::Cache;
-use cargo_metadata::MetadataCommand;
 use config_file::ConfigFile;
 use eyre::WrapErr;
 use http::StatusCode;
-use kinetics_parser::Package;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-
-// Workspace definition for a project.
-//
-// A non-workspace project would have one member
-// with it's path identical to workspace_root.
-//
-// Otherwise it's a real workspace.
-#[derive(Debug, Default, Clone)]
-pub struct Workspace {
-    pub root_path: PathBuf,
-    pub packages: Vec<Package>,
-}
-
-impl Workspace {
-    pub fn from_path(path: &Path) -> eyre::Result<Self> {
-        let metadata = MetadataCommand::new().current_dir(path).exec()?;
-
-        // Validate workspace rules for kinetics:
-        // 1. Workspace as a single kinetics project:
-        //  - the workspace root MUST contain kinetics.toml;
-        //  - workspace members cannot have kinetics.toml - throw an error.
-        // 2. Standalone kinetics project:
-        //  - the project can contain kinetics.toml;
-        //  - for the name fall back to Cargo.toml.
-        let root_config = metadata.workspace_root.join("kinetics.toml");
-        let members_configs: Vec<_> = metadata
-            .workspace_members
-            .iter()
-            .filter_map(|member| {
-                metadata
-                    .packages
-                    .iter()
-                    .find(|pkg| pkg.id == *member)
-                    .and_then(|pkg| pkg.manifest_path.parent())
-                    .and_then(|dir| {
-                        let config = dir.join("kinetics.toml");
-                        if config.exists() {
-                            Some(config)
-                        } else {
-                            None
-                        }
-                    })
-            })
-            .collect();
-
-        if root_config.exists() && !members_configs.is_empty() {
-            return Err(eyre::eyre!("Workspace is not allowed to have `kinetics.toml` within its root and within its members at the same time."));
-        }
-
-        Ok(Self {
-            packages: metadata
-                .workspace_members
-                .into_iter()
-                .filter_map(|member| {
-                    metadata
-                        .packages
-                        .iter()
-                        .find(|pkg| pkg.id == member)
-                        .and_then(|pkg| {
-                            Some(Package {
-                                name: ConfigFile::cargo_toml_name(
-                                    pkg.manifest_path.parent()?.as_std_path(),
-                                )
-                                .ok()?,
-                                relative_path: pkg
-                                    .manifest_path
-                                    .strip_prefix(&metadata.workspace_root)
-                                    .ok()?
-                                    .parent()? // Remove filename and keep only the dir name.
-                                    .into(),
-                            })
-                        })
-                })
-                .collect(),
-            root_path: metadata.workspace_root.into_std_path_buf(),
-        })
-    }
-}
+use std::path::PathBuf;
 
 /// Managing user's project
 ///
@@ -316,5 +238,18 @@ impl Project {
     /// No need to store it in Project props, it's not going to be loaded frequently
     pub fn environment(&self) -> HashMap<String, String> {
         Envs::load()
+    }
+}
+
+impl From<ProjectInfo> for Project {
+    fn from(value: ProjectInfo) -> Self {
+        Self {
+            path: PathBuf::new(),
+            workspace: Workspace::default(),
+            name: value.name,
+            url: value.url,
+            kvdb: value.kvdb,
+            observability: None,
+        }
     }
 }
